Html structure
Built html structure with little css.
In this case we only have 3 buttons which is why I preferred this approach then event delegation on the whole section.
Create element helper function is handy here to create a li tag.
Currently this is not storing in local storage correctly.
The only features are adding to the list.