functions to insert  user input to the destination list
functions to insert  user input to the destination list, that uses the element creator fucntion and dom appender function to append it to th dom.

When the user clicks on one of the add-task buttons, a new task will be added to the respective list. The task content will be taken from the respective input field.

Co-Authored-By: Cyber4sPopo <69596366+cyber4spopo@users.noreply.github.com>