Tomer Yacobi Final Project.
A project that displays a tasks list, the tasks divide into three sections to-do, in-progress , done.
the tasks are saved into the browser local storage, and a dedicated api.
a client can customize the tasks list move tasks, edit them, save, and load tasks.
this project was made by tomer yacobi.