MASSIVE BUG FIXES!!!
had so many bugs i had to kinda start over... still workin on it... gettin it fixed by tommmorrow 27/09/21.

this is what need to be done as welll which worked and stopped working after applied the API:

(1) make save and load (&clear?) buttons so user can load throguh the api content as well.

(2) whenever smoething get deleted or the text of list gets updated (by the 2 clicks event thing..) then update it in localstorage as well

(3) whenever page reloads... make event listeners work as well. cuz when reloading the events disattached from the tasks for some reason

(4) make DRAG & DROP task listener. so the tasks can be dragged to other list (AND BE UPDATED IN THE LOCAL STORAGE)

(5) ONE OF THE MOST IMPO ONES: I AM TRYNA LOAD THE DATE AND TIME OF  CREATION OF THE ITEMS AS WELL TO THE LIST AND TO THE DATA AND API ... I HAD IT BEFORE BUT NOW IT SEEMS OT BE GONE AND ALSO NOT STORED IN API AND LOCAL STORAGE!

thank you.

Co-Authored-By: Cyber4sPopo <69596366+cyber4spopo@users.noreply.github.com>