Fixed JSON handling
To my mistake, I thought the API needed  the value of its tasks property to be stored as a string (like in local storage).
Hence, I sent it over as a string and did some ugly string replacement when loading it back to the browser.
Only later I understood the proper way of doing this, although the former way still worked it was certainly NOT ideal.
Plus, it deleted '\' from card text if there was any.